+++ js/add-user.js
@@ -1,6 +1,8 @@
-import { MAX_USERS } from './data.js';
-import { createUser } from './create-user.js';
 
-const addUser = () => Array.from({ length: MAX_USERS }, createUser);
+import { createPhoto } from './create-photo.js';
+
+const MAX_USERS = 25;
+
+const addUser = () => Array.from({ length: MAX_USERS }, createPhoto);
 
 export { addUser };
